third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-removechild-and-blur-event-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: test.createShadowRoot is not a function
Check focused shadow node dispatchs blur event on removeChild of host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
